BIG refactoring, add far more tokens in the different ASTs and AST generic
Description: sgrep sometimes fails with an OCaml exception about an empty
list of tokens (either FakeTok error or min_max_ii error) because
a metavariable may match something (an expr or stmt) that does not contain
any token information. This in turns mean we can not display location
information for this matched thing.
This diff should help fix
https://github.com/returntocorp/sgrep/issues/73
and
https://github.com/returntocorp/sgrep/issues/52

because it adds lots of token information in the different ASTs.
I originally skipped those tokens because I wanted an AST, not a CST,
but in the ends it bites back and we want many of those tokens
(at least enough token so we never get the error above).

See ast_generic.ml for the rational.

test plan:
make test