* commons/ocaml.ml: call callback in map_of_ref.
Before this diff, any ref elements in an AST was just copied-as-is
in a map, e.g., in the map_ast.ml which converts one AST generic into another,
especially the one that abstract all token info so that we can use
A = B to compare 2 big AST generics.
That means that the concrete token location buried behind a ref
(e.g., the resolved name for a Global), was not visited/converted and
so was not abstracted.

This should fix issue https://github.com/returntocorp/sgrep/issues/92

Test plan:
Follow instructions in issue above; now sgrep reports far more matches.