Use parsing JS hacks also for the sgrep pattern, fix arrow parsing
Some arrows using parenthesis, as in (a) => { ... } require
some parsing hacks to convert the open parenthesis in a special token.
I was not performing this hack for the sgrep pattern, which prevented
to parse certain arrows in a sgrep pattern.

This should help fix https://github.com/returntocorp/sgrep/issues/267

Test plan:
test files for sgrep included.