Improve fstrings (#108)
* Properly lex python fstrings with spaces in interpolants

Previously, Python would fail to lex Python fstring interpolants if
trailing space appeared in the interpolant. This was because the lexer
state was incorrectly changed by the whitespace lexing, stomping on the
correct "fstring" state mode.

Here, I only apply this state change if the origin state is the assumed
state (in this case, _token -> token).

For future debugging, I've also added Lexer.pr_state, which can be used
to print the entire lexer state stack.

I've also added test cases for these fstrings.

* Treat Python F"..." as an f-string

Previously, a leading F was treated as a generic literal string prefix.
However, it signifies the introduction of a python fstring in Python,
and should be treated as such.

Updated fstring2.py to match.