Added script to check menhir conflicts and modified Makefile for js
Addresses issue 800: ensure the number of grammar conflicts does not grow in pfff. Wrote a script that takes in a menhir error file, expected number of errors, and certain file paths, and errors if this number is greater than a given constant. Modified Makefile for lang_js to run this script.

Verified by running make in lang_js/parsing as well as top level, with expected conflicts 1 (js has 2 conflicts).