Support use of interface{} inside if/for/..., parsing hack
The Go grammar is not LALR(1). To manage some ambiguity we need
to use some parsing hack to find the opening brace of an if/for/switch/select.
This add a new heuristic to not treat the { of interface as the
opening brace.

This fixes https://github.com/returntocorp/semgrep/issues/829

Test plan:
test file included