Made initial changes to implement param/local type propagation
Modified Naming_AST.ml to support parameter and local type propagation. Co-opted the existing field id_type in id_info, and modified the name to be a record with entname and entype, so that the type information is stored in scopes.

Tested by running ./pfff -typing_generic tests/GENERIC/typing/[test.java] and checking manually.