Add support for arbitrary decorator matching in python
We want to support matching an arbitrary number of decorators, so we can
search for a given decorator among many.

E.g. match

@a
@b
@c
def function():
    pass

With a search on only @c.

To enable this, I add syntax: @...

This syntax will be used to match 0 or more decorators, so the example
can be matched with the pattern:

@...
@b
@...
def $FUNCTION(...):
    ...

This commit enables parsing of `@...`.

A test case is added in decorator-pattern.py.