Improve js array destructuring
We were failing to parse a number of array destructuring cases; viz. the
empty array, and arrays with a single leading comma.

Here I update destructuring parsing to handle these two cases.

We were also successfully parsing destructuring with a '...<pattern>'
not in the last position, which is disallowed according to ES7 syntax.
I have also updated parsing to only parse this in last position.

Test cases are updated, and I also added a negative test case.