Added support for typed metavariables in method calls.
In primary, added a case that matches typed metavariables. In addition, since currently I believe we do not have support for propagating the values of types as metavariables, added logic that raises fail on a pattern like ($T $A) (this may need to be changed to a ParseError later).

Tested against the cases

$T $B; (should pass)
($T $B); (should fail)
(String $B); (should pass)
(String $B).func)(); (should fail)