Parse rest param in object destructuring (#143)
Here we transpile:

  var {...a} = b;

into

  var a = <spread>(b);

This is a bit of a hack, as really the rest param here is equivalent to
ommitting unselected keys. However, there's no built-in within
ECMAScript to perform this omission beyond the object rest param.

Note that this will cause us to confuse the rest parameter in these two
cases:

  var {...a} = b;
  var {x, ...a} = b;

I'm leaving this as an existing wart for now.

The "spread_params.js" test case is updated (although this is a "rest",
rather than a "spread" parameter, with inverse functionality).