Fix regression in semgrep on disabling naming of OT_Expr
This is related to https://github.com/returntocorp/semgrep/issues/923

In my previous fix I've disabled the recursion on OtherType to avoid
creating cycles in the AST but doing so I've introduced some regressions
in semgrep (in tests/python/misc_regression1.py) because we were not
resolving correctly aliases anymore.

This fixes the fix.

Test plan:
+ /home/pad/pfff/pfff -dump_ast -naming_generic python/misc_regression1.py
now shows the right id_resolved

and we still do not go into infinite loop