Finish unicode hack by Martin
Integrate in build system and add explicit (option) parameter
to enable the unicode hack.

Test plan:
a few test files included.

I've also tested the parser before/after on huge corpus
of Java, Go, JS, and Python. It fixes a few parse errors.
There is a slow down but it's not too big (and it can be optimized
I think).

Java: on 63063 files in 2m14s 98.61% to 98.62% in 2m50
Go: on 38974 files in 1ms 99.64% to 99.77% in 2m41
Python: on 14040 files in 14s 99.95% to 99.97% in 22s
Javascript: on 12788 files in 1m2 91.28% to 92.03% in 1m20s