Parse trailing whitespace in Python files
For whatever reason, the Python interpreter allows you to leave dangling
whitespace after the last statement in a Python file, and this does not
actually get interpreted as indentation.

I couldn't figure out exactly what the lexer was doing here, but it
parses trailing whitespace (all on one line) as:

  INDENT NEWLINE DEDENT NEWLINE EOF

Here I just match against this magic code and drop it at the end of
a file.