Parse python2 implicit array in for comprehension
In Python 2, the following construct is allowed:

  [x for x in 1, 2]

This commit adds parsing support for this.

Note that adding this is a tad cumbersome, as Python 3 allows:

  foo(x for x in bar, baz)

Note that in the first case, the comma separates elements within the
"in" clause; in the second case, the comma separates the entire for
comprehension from the second argument.

This means that we have to use completely different grammar rules inside
list atoms from those inside function calls.

I've added test cases for both syntaxes here.