Eliminate a Python S/R conflict
We had production rules for sgrep patterns for both a single expression,
and and expression statement. These conflict trivially.

Note that the production pattern for single expressions was being
eliminated by YACC anyway, so here I just remove it. This was proved via
testing with the included tests/python/parsing/test.sgrep file.