Create generic AST from nested JS patterns
Previously, our transpilation step (necessary for creating a generic
AST) for JavaScript could not handle nested array patterns.

Note that JavaScript does not support nested object patterns (instead,
one would assign to the right-hand-side of a property pattern), so we
fail hard in this case.

After this commit, patterns like:

  function foo([a, [b, {c}]]) { ... }

Will be transpiled to:

  function foo(arg0) {
    let a = arg0[0];
    let b = arg0[1][0];
    let c = arg0[1][1].c;
  }