Do not transpile ParamPattern for Javascript
It is better to stay close to the original code and not do too much
magic. I was transpiling for the abstract interpreter to simplify
things but for semgrep better not too. We can still reuse the
transpiling code for the IL (and it can be generic!)

Test plan:
make
also make reinstall-libs and doing make test in semgrep seems to still work!