* commons_core/parallel.ml: fix invoke in case of broken pipe (#283)
This was very subtle to debug, but when running a command like:

  semgrep -l js --timeout 0.5 -e 'FOO' tests/PERF/timeout.js

or even the simpler

   /home/pad/github/semgrep/semgrep-core/_build/default/bin/Main.exe -timeout 0.5 -lang js -e FOO tests/PERF/timeout.js -debug -profile

would actually takes 12s to finish! not 0.5s.

Indeed, this is because we currently use Parallel.invoke to run the
tree-sitter parser in its own process (because it segfault or leak memory).
In the case of a short timeout, the timeout is triggered in the
parent process (alarm or timers are not inherited through fork) which
would then finish and close the pipe to the child process. In that case
the child would get an exn that would bubble up and not trigger the exit call,
which would then in Parse_code be captured and would trigger the parsing
of the javascript pfff parser (which takes 12s to finish on timeout.js).

Subtle.

test plan:
both

/home/pad/github/semgrep/semgrep-core/_build/default/bin/Main.exe -timeout 0.5 -lang js -e FOO tests/PERF/timeout.js -debug -profile

and
  semgrep -l js --timeout 0.5 -e 'FOO' tests/PERF/timeout.js

finish quickly