Add WithUsingResource to generic AST (#346)
* Add WithUsingResource to generic AST

Obtain a resource using some expression or variable declaration, and dispose of
that value at the end of the block.

The first statement would be a ExprStmt or DefStmt to acquire some resource,
and the second statement would be the block where the resource is used.

E.g. in python:
```python
with fopen(file) as fp:
	...
```

E.g. in C#:
```csharp
using (var reader = new StringReader(manyLines))
{ ... }
```

E.g. in Java:
```java
try (Foo f = new Foo())
{ ... }
```

Resolves #342

* Add newscope: comment to block of WithUsingResource

The body of a using-statement is a new scope.

* Add WithUsingResource

For C#'s using statement, or Java's try-with-resource.

Treat it as two consecutive statements in most places.

* Fix indention