Moving AST_generic.ml out of h_program-lang into its own lang_GENERIC_base dir
h_program-lang is getting messy, and there are more and more files
related to the generic AST in this dir. Better to split.
Moreover, I want to have a generic code highlighter, which means
I need a highlight_generic.ml, which would have to be in h_program-lang,
and this file requires Naming_AST.ml, which would be another file again,
so better to split.

test plan:
make
make test