[JS] use '. ...' for method chaining to remove ambiguity
We can't simply use "..." for method chaining in semgrep pattern as in
  $o.foo() ... .bar()
because there is an ambiguity with ASI for patterns like
  $o.foo()
  ...
so we have to force the extra "." before.

This should help https://github.com/returntocorp/semgrep-rules/issues/1055
This changes https://github.com/returntocorp/semgrep/issues/2208

test plan:
see related PR in semgrep.