Use stmt instead of stmt list for body of for/do/while
Closes #2453

In php, the cst uses a stmt to represent the body of loops, as does the generic ast, but the ast uses a stmt list. This PR changes it to go directly through a statement. Additionally, blocks remain blocks, so that single statement blocks remain blocks.

Test plan: run make test in semgrep, with the added files in the associated PR