Add Commons utility to copy FIFOs to temp files
UNIX FIFOs can only be read once. For programs that require repeated
reads to a queue (e.g. Semgrep, which reads once to parse, and once to
report matches), we must therefore copy the contents of these queues.

Here I adopt an API that returns the list of all files, FIFO or not,
that the program should open. In order to support temp-file cleanup,
I also provide a list of all created temp files.