Fix the license from GPL to LGPL in many parse_xxx.ml
All those files should have been LGPL in the first place.
If you git blame license.txt at the top you'll see pfff was
meant to be LGPL from the start in 2010 when it was open sourced.
There was a few files with GPL headers because many of them
started from a copy-pasted template that had the wrong header.