Use AST enums instead of CST enums when building AST
Class modifiers and cast types would previously have the CST enum in the AST. Now we explicitly define AST enums for those values, and map from CST values to AST values. This reduces the dependency on the CST and makes it easier to switch to another parser (i.e. tree-sitter-php).