ruby: Fix translation of foo[bar]
It was translated as `foo.](bar)`, where `foo.]` had the same range as
`foo[:bar]` itself. This caused problems with taint-mode and field
sensitivity. If `foo[bar]` is tainted then `foo.]` is tainted by
side-effect, and then `foo[other]` ends up being tainted too.

We now translate it as `foo.[(bar)` instead... we use `[` and `]` tokens
instead of the fake `(` and `)` ones so the range of the Call node is
the correct one.

Helps PA-2087

Fixes: 129c24294a8 ("ruby: Fix translation of foo[:bar]")

test plan:
Dump the AST of `foo[:bar]` and check it